created nav bar.  migrated signup/signin/signout to nav bar
