+++ public/client/Components/App.jsx
@@ -3,10 +3,11 @@
 import React, { useEffect, useState } from 'react';
 import fetch from 'isomorphic-fetch';
 import Map from './Map.jsx';
-import LogIn from './LogIn.jsx';
-import Welcome from './Welcome.jsx';
+// import LogIn from './LogIn.jsx';
+// import Welcome from './Welcome.jsx';
 import FavoriteList from './FavoriteList.jsx';
 import NewsFeed from './NewsFeed.jsx';
+import NavBar from './NavBar.jsx';
 
 function App() {
   const [currentFavorites, setFavorites] = useState({});
@@ -15,73 +16,6 @@ function App() {
   const [currentUser, changeUser] = useState(null);
   const [currentCountryClick, setCurrentCountryClick] = useState(null);
   const [posts, setPosts] = useState([]);
-
-  const loginButton = (e) => {
-    const username = document.querySelector('#username');
-    const password = document.querySelector('#password');
-
-    if (username.value === '' || password.value === '') {
-      const result = 'Please fill out the username and password fields to log in.';
-      changeAttempt(result);
-    } else {
-      const user = {
-        username: username.value,
-        password: password.value,
-      };
-      fetch('/api/login', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify(user),
-
-      })
-        .then((res) => res.json())
-        .then((data) => {
-          if (!Array.isArray(data)) throw Error('wrong');
-          if (Array.isArray(data)) {
-            setFavorites({});
-            const favoritesObj = {};
-            data.forEach((elem) => {
-              favoritesObj[elem.title] = elem.link;
-            });
-            setFavorites(favoritesObj);
-            changeUser(username.value);
-            changeLoginStatus(true);
-          }
-        })
-        .catch((err) => changeAttempt('Incorrect username or password!'));
-    }
-  };
-
-  const signUp = (e) => {
-    const username = document.querySelector('#username');
-    const password = document.querySelector('#password');
-
-    if (username.value === '' || password.value === '') {
-      const result = 'Please fill out the username and password fields to sign up.';
-      changeAttempt(result);
-    } else if (password.value.length < 5) {
-      const result = 'Please create a password longer than 5 characters';
-      changeAttempt(result);
-    } else {
-      const user = {
-        username: username.value,
-        password: password.value,
-      };
-      fetch('/api/signup', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify(user),
-      })
-        .then((res) => {
-          if (res.status === 200) {
-            changeLoginStatus(true);
-            changeUser(username.value);
-          }
-        })
-
-        .catch((err) => console.log(err));
-    }
-  };
 
   const getPosts = (countryName) => {
     setTimeout(async () => {
@@ -118,41 +52,40 @@ function App() {
     });
   };
 
-  const signOut = () => {
-    console.log('test');
-    changeLoginStatus(false);
-    changeAttempt(null);
-    setFavorites({});
-    changeUser(null);
-    setCurrentCountryClick(null);
-    setPosts([]);
-  };
-
   return (
-    <div className="wrapper">
-      {!loginStatus
-        ? <LogIn loginButton={loginButton} signUp={signUp} loginAttempt={loginAttempt} />
-        : [<Welcome key={1} currentUser={currentUser} signOut={signOut} />,,
-        ]}
-      <Map
+    <>
+      <NavBar
+        setFavorites={setFavorites}
+        loginStatus={loginStatus}
+        changeLoginStatus={changeLoginStatus}
+        loginAttempt={loginAttempt}
+        changeAttempt={changeAttempt}
+        currentUser={currentUser}
+        changeUser={changeUser}
         setCurrentCountryClick={setCurrentCountryClick}
         setPosts={setPosts}
-        getPosts={getPosts}
       />
-      <NewsFeed
-        currentCountryClick={currentCountryClick}
-        posts={posts}
-        currentFavorites={currentFavorites}
-        setFavorites={setFavorites}
-        addFavorite={addFavorite}
-        deleteFavorite={deleteFavorite}
-      />
+      <div className="wrapper">
+        <Map
+          setCurrentCountryClick={setCurrentCountryClick}
+          setPosts={setPosts}
+          getPosts={getPosts}
+        />
+        <NewsFeed
+          currentCountryClick={currentCountryClick}
+          posts={posts}
+          currentFavorites={currentFavorites}
+          setFavorites={setFavorites}
+          addFavorite={addFavorite}
+          deleteFavorite={deleteFavorite}
+        />
 
-      <FavoriteList
-        currentFavorites={currentFavorites}
-        deleteFavorite={deleteFavorite}
-      />
-    </div>
+        <FavoriteList
+          currentFavorites={currentFavorites}
+          deleteFavorite={deleteFavorite}
+        />
+      </div>
+    </>
   );
 }
