+++ public/client/Components/LogIn.jsx
@@ -1,29 +1,20 @@
 /* eslint-disable react/button-has-type */
 /* eslint-disable react/no-array-index-key */
 import React from 'react';
-import PropTypes from 'prop-types';
+// import PropTypes from 'prop-types';
 
 function LogIn(props) {
   const { loginButton, signUp, loginAttempt } = props;
   return (
-    <div>
-      Log In
-      <div id="inputButtonWrapper">
-        <input name="username" placeholder="username" id="username" autoComplete="off" />
-        <input name="password" placeholder="password" id="password" autoComplete="off" type="password" />
-        <div id="buttonsDiv">
-          <button onClick={loginButton} value="Log-In"> Log In </button>
-          <button onClick={signUp} value="Sign-Up"> Sign-Up </button>
-        </div>
-        <div />
-        <div id="loginAttemptMessage">
-          {loginAttempt}
-        </div>
-      </div>
-      <div>
-        <div className="g-signin2" data-onsuccess="onSignIn">Log in with Google</div>
-      </div>
-    </div>
+    <>
+      <div className="g-signin2" data-onsuccess="onSignIn">Log in with Google</div>
+      <button onClick={signUp} value="Sign-Up"> Sign-Up </button>
+      <button onClick={loginButton} value="Log-In"> Log In </button>
+      <input name="password" placeholder="password" id="password" autoComplete="off" type="password" />
+      <input name="username" placeholder="username" id="username" autoComplete="off" />
+      <div id="loginAttemptMessage">{loginAttempt}</div>
+      <h1>World Wide News</h1>
+    </>
   );
 }
