+++ public/index.html
@@ -11,7 +11,7 @@
 
 </head>
 <body>
-  <h1 class = "header">World Wide News</h1>
+  <!-- <h1 class = "header">World Wide News</h1> -->
   <div id = "root"></div>
 
   <!-- script for webpack and webpack-dev-server challenge -->
